refactor(logger): use class for consistency
